Split up `participantsRouter` file
- Split up the `participantsRouter.ts` file based on domain.
- Logically group the routes, ensuring the correct middleware still gets run
