+++ src/web/components/ParticipantManagement/ParticipantRequestItem.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 
 import { ApiRoleDTO } from '../../../api/entities/ApiRole';
 import { ParticipantTypeDTO } from '../../../api/entities/ParticipantType';
-import { ParticipantRequestDTO } from '../../../api/routers/participants/participantsRouter';
+import { ParticipantRequestDTO } from '../../../api/services/participantsService';
 import { ParticipantApprovalFormDetails } from '../../services/participant';
 import { Dialog } from '../Core/Dialog/Dialog';
 import { InlineMessage } from '../Core/InlineMessages/InlineMessage';
